resume-search-mcp/requirements.txt
# requirements.txt
mcp[cli]
httpx
python-dotenv